+++ lumenairy/__init__.py
@@ -1016,7 +1016,7 @@ load_zemax_prescription_txt = _deprecated_alias(
     version_removed='5.0',
 )
 
-__version__ = "5.24.0"
+__version__ = "5.24.1"
 
 #
 # __all__ is grouped by user-journey tier: